Removed interface Environment
